% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GD.functions.R
\name{genome.doub.timing}
\alias{genome.doub.timing}
\title{genome.doub.timing}
\usage{
genome.doub.timing(sample, seg.mat.minor, seg.mat.copy, loss.level = 3)
}
\arguments{
\item{sample}{TCGA sample identifier}

\item{seg.mat.minor}{segmented minor allele copy number matrix. segmented matrix with minor allele (e.g. seg.mat.LOH). columns should be as follows: c("Sample","Chrom","Start","End","Num.probes","val")}

\item{seg.mat.copy}{segmented total copy number matrix. segmented total copy number matrix (columns as above)}

\item{loss.level}{The cnv loss types to use. 2 or 3. If loss.level == 2, genotypes, AA and AB are used, if loss.level == 3, genotypes: AA, AB, AAB and AAA are used.}
}
\value{
the function returns a GD.timing dataframe and the information of arm levels.
}
\description{
Function to assess the timing of GD relative to the genome loss. Some genotypes are happened before GD (AA or AAA) whereas the other are after GD (AAB and AB)
}
\details{
According to the code of clonality.estimation(), "val" means "cn". the function uses the limma weighted.median function
}
\examples{
genome.doub.timing(sample, seg.mat.minor, seg.mat.copy)

}
