% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.functions.r
\name{identify.mut.copy.number.ascat}
\alias{identify.mut.copy.number.ascat}
\title{identify.mut.copy.number.ascat}
\usage{
\method{identify}{mut.copy.number.ascat}(x, sub.mat.mut, sub.mat.copy)
}
\description{
identify.mut.copy.number.ascat
}
