% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.functions.r
\name{plot.EarlyOrLate}
\alias{plot.EarlyOrLate}
\title{plot.EarlyOrLate}
\usage{
\method{plot}{EarlyOrLate}(
  seg.mat.patient,
  TCGA.earlyLate,
  TCGA.purity,
  TCGA.barcode,
  max.cpn = 5,
  min.probes = 10,
  sub.clonal = 1
)
}
\description{
plot.EarlyOrLate
}
