% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.functions.r
\name{plot.TCGA.sample.ccfs}
\alias{plot.TCGA.sample.ccfs}
\title{plot.TCGA.sample.ccfs}
\usage{
\method{plot}{TCGA.sample.ccfs}(TCGA.earlyLate, clonal.cut.off = 1, spec.genes = NULL)
}
\description{
plot.TCGA.sample.ccfs
}
